requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scipy>=1.7.0
openpyxl>=3.0.0
PyQt5>=5.15.0
matplotlib>=3.4.0
pyyaml>=6.0
pytest>=6.2.0
PyInstaller>=4.5.0 